fix code

+++ code/05_03/05_03.cpp
@@ -2,7 +2,7 @@
 using namespace std;
 
 int main(){
-    long long n;
+    int n;
     cin >> n;
     int arr[n];
     for (int i = 0; i < n; i++){
